initial work on updatereqcon refactor
